refactor(unittests): move unit tests to CaseExtensions.UnitTests project
