+++ src/CaseExtensions/CaseExtensions.csproj
@@ -1,4 +1,4 @@
-<Project Sdk="Microsoft.NET.Sdk">
+﻿<Project Sdk="Microsoft.NET.Sdk">
 
     <PropertyGroup>
         <!-- <TargetFrameworks>netstandard1.3;netstandard2.0;net8.0</TargetFrameworks> -->
@@ -7,22 +7,4 @@
         <OutputType>Library</OutputType>
     </PropertyGroup>
 
-    <!-- Test -->
-    <PropertyGroup Condition="'$(Configuration)' == 'Debug'">
-        <CoverletOutput>$(CoverletOutput)CaseExtensions</CoverletOutput>
-    </PropertyGroup>
-
-    <ItemGroup>
-        <Compile Remove="$(TestFiles)" Condition="'$(Configuration)' != 'Debug'"/>
-    </ItemGroup>
-
-    <ItemGroup Condition="'$(Configuration)' == 'Debug'">
-        <PackageReference Include="coverlet.msbuild" Version="2.9.0" PrivateAssets="All" />
-        <PackageReference Include="FluentAssertions" Version="5.10.3" PrivateAssets="All" />
-        <PackageReference Include="xunit" Version="2.4.1" PrivateAssets="All" />
-    </ItemGroup>
-
-    <ItemGroup Condition="'$(TargetFramework)' == 'net5.0'">
-        <PackageReference Include="xunit.runner.visualstudio" Version="2.4.1" PrivateAssets="All" />
-    </ItemGroup>
 </Project>
